requirements.txt
Django
django_environ
gunicorn
pip
whitenoise
psycopg2
psycopg2-binary